+++ src/main/java/domain/expression/operator/Operator.java
@@ -1,4 +1,4 @@
-package domain.calculator.operator;
+package domain.expression.operator;
 
 public abstract class Operator {
